common: move curl to crypto
