+++ common/crypto/curl-p/const.c
@@ -7,8 +7,8 @@
 
 #include <stdlib.h>
 
-#include "common/curl-p/const.h"
-#include "common/curl-p/indices.h"
+#include "common/crypto/curl-p/const.h"
+#include "common/crypto/curl-p/indices.h"
 
 size_t const CURL_INDEX[STATE_LENGTH + 1] = {__INDEX_TABLE};
 trit_t const TRUTH_TABLE[11] = {__TRUTH_TABLE};
