+++ common/crypto/curl-p/digest.c
@@ -7,7 +7,7 @@
 
 #include <string.h>
 
-#include "common/curl-p/digest.h"
+#include "common/crypto/curl-p/digest.h"
 #include "common/trinary/trit_tryte.h"
 
 void curl_digest(trit_t const* const trits, size_t const len, trit_t* const out, Curl* const curl) {
