+++ common/crypto/curl-p/hamming.h
@@ -8,8 +8,8 @@
 #ifndef __COMMON_CURL_P_HAMMING_H_
 #define __COMMON_CURL_P_HAMMING_H_
 
-#include "common/curl-p/pearl_diver.h"
-#include "common/curl-p/trit.h"
+#include "common/crypto/curl-p/pearl_diver.h"
+#include "common/crypto/curl-p/trit.h"
 
 #ifdef __cplusplus
 extern "C" {
