+++ common/crypto/curl-p/pearl_diver.c
@@ -9,8 +9,8 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "common/curl-p/ptrit.h"
-#include "common/curl-p/search.h"
+#include "common/crypto/curl-p/ptrit.h"
+#include "common/crypto/curl-p/search.h"
 #include "common/trinary/ptrit_incr.h"
 #include "common/trinary/trit_ptrit.h"
 #include "utils/handles/rw_lock.h"
