+++ common/crypto/curl-p/trit.h
@@ -10,7 +10,7 @@
 
 #include <string.h>
 
-#include "common/curl-p/const.h"
+#include "common/crypto/curl-p/const.h"
 #include "common/trinary/trits.h"
 
 #ifdef __cplusplus
