+++ common/crypto/kerl/hash.c
@@ -6,6 +6,7 @@
  */
 
 #include "common/crypto/kerl/hash.h"
+#include "common/defs.h"
 #include "common/trinary/tryte.h"
 
 void kerl_hash(const trit_t* const trits, size_t len, trit_t* out, Kerl* kerl) {
