+++ common/sign/v1/tests/BUILD
@@ -4,7 +4,6 @@ cc_test(
         "test_iss_curl.c",
     ],
     deps = [
-        "//common/curl-p:trit",
         "//common/sign/v1:iss_curl",
         "//common/trinary:trit_tryte",
         "@unity",
@@ -17,7 +16,6 @@ cc_test(
         "test_iss_kerl.c",
     ],
     deps = [
-        "//common/curl-p:trit",
         "//common/sign/v1:iss_kerl",
         "//common/trinary:trit_tryte",
         "@unity",
