+++ gossip/components/BUILD
@@ -40,7 +40,7 @@ cc_library(
     srcs = ["processor.c"],
     deps = [
         ":processor_shared",
-        "//common/curl-p:ptrit",
+        "//common/crypto/curl-p:ptrit",
         "//common/trinary:trit_ptrit",
         "//consensus/milestone_tracker",
         "//consensus/transaction_solidifier",
