+++ mam/prototype/mam.c
@@ -8,7 +8,7 @@
 #include <stdio.h>
 #include <string.h>
 
-#include "common/curl-p/hamming.h"
+#include "common/crypto/curl-p/hamming.h"
 #include "common/sign/v2/iss_curl.h"
 #include "common/trinary/trit_long.h"
 #include "mam/prototype/mask.h"
